+++ support.py
@@ -2,9 +2,21 @@
 import sublime
 import sublime_plugin
 import textwrap
+import webbrowser
+import re
 
 __version__ = "3.1.0"
 __pc_name__ = 'RegReplace'
+
+CSS = '''
+div.reg-replace { padding: 10px; margin: 0; }
+.reg-replace h1, .reg-replace h2, .reg-replace h3,
+.reg-replace h4, .reg-replace h5, .reg-replace h6 {
+    {{'.string'|css}}
+}
+.reg-replace blockquote { {{'.comment'|css}} }
+.reg-replace a { text-decoration: none; }
+'''
 
 
 def list2string(obj):
@@ -104,3 +116,95 @@ class RegReplaceSupportInfoCommand(sublime_plugin.ApplicationCommand):
 
         sublime.message_dialog(msg + '\nInfo has been copied to the clipboard.')
         sublime.set_clipboard(msg)
+
+class RegReplaceOpenSiteCommand(sublime_plugin.ApplicationCommand):
+    """Open site links."""
+
+    def run(self, url):
+        """Open the url."""
+
+        webbrowser.open_new_tab(url)
+
+
+class RegReplaceDocCommand(sublime_plugin.WindowCommand):
+    """Open doc page."""
+
+    re_pkgs = re.compile(r'^Packages')
+
+    def on_navigate(self, href):
+        """Handle links."""
+
+        if href.startswith('sub://Packages'):
+            sublime.run_command('open_file', {"file": self.re_pkgs.sub('${packages}', href[6:])})
+        else:
+            webbrowser.open_new_tab(href)
+
+    def run(self, page):
+        """Open page."""
+
+        try:
+            import mdpopups
+            has_phantom_support = (mdpopups.version() >= (1, 10, 0)) and (int(sublime.version()) >= 3118)
+        except Exception:
+            has_phantom_support = False
+
+        if not has_phantom_support:
+            sublime.run_command('open_file', {"file": page})
+        else:
+            text = sublime.load_resource(page.replace('${packages}', 'Packages'))
+            view = self.window.new_file()
+            view.set_name('RegReplace - Quick Start')
+            view.settings().set('gutter', False)
+            view.settings().set('word_wrap', False)
+            if has_phantom_support:
+                mdpopups.add_phantom(
+                    view,
+                    'quickstart',
+                    sublime.Region(0),
+                    text,
+                    sublime.LAYOUT_INLINE,
+                    css=CSS,
+                    wrapper_class="reg-replace",
+                    on_navigate=self.on_navigate
+                )
+            else:
+                view.run_command('insert', {"characters": text})
+            view.set_read_only(True)
+            view.set_scratch(True)
+
+
+class RegReplaceChangesCommand(sublime_plugin.WindowCommand):
+    """Changelog command."""
+
+    def run(self):
+        """Show the changelog in a new view."""
+        try:
+            import mdpopups
+            has_phantom_support = (mdpopups.version() >= (1, 10, 0)) and (int(sublime.version()) >= 3118)
+        except Exception:
+            has_phantom_support = False
+
+        text = sublime.load_resource('Packages/RegReplace/CHANGES.md')
+        view = self.window.new_file()
+        view.set_name('RegReplace - Changelog')
+        view.settings().set('gutter', False)
+        view.settings().set('word_wrap', False)
+        if has_phantom_support:
+            mdpopups.add_phantom(
+                view,
+                'changelog',
+                sublime.Region(0),
+                text,
+                sublime.LAYOUT_INLINE,
+                wrapper_class="reg-replace",
+                css=CSS,
+                on_navigate=self.on_navigate
+            )
+        else:
+            view.run_command('insert', {"characters": text})
+        view.set_read_only(True)
+        view.set_scratch(True)
+
+    def on_navigate(self, href):
+        """Open links."""
+        webbrowser.open_new_tab(href)
